# AR VirtualAttribute queries + introspection