--- a/app/controllers/api/base_controller/renderer.rb
+++ b/app/controllers/api/base_controller/renderer.rb
@@ -575,13 +575,20 @@ module Api
       end
 
       def determine_include_for_find(klass)
-        virtual_attributes_for(klass) do |type, attr_name, attr_base|
+        attrs = virtual_attributes_for(klass) do |type, attr_name, attr_base|
           next if attr_base.blank?
-          next if attr_base.include?(".") # FIXME: Allow nested relations
           next if virtual_attribute_accessor(type, attr_name)
           next if Rbac::Filterer::CLASSES_THAT_PARTICIPATE_IN_RBAC.include?(attr_base)
 
           attr_base
+        end
+
+        # Handle nested relationships and convert to a hash
+        if attrs
+          attrs.each_with_object({}) do |key, include_for_find|
+            nested = include_for_find
+            key.split(".").each { |k| nested = nested[k] ||= {} }
+          end
         end
       end