@NickLaMuro 2020-08-31
@miq-bot assign @jrafanie 
@miq-bot add_labels performance, jansa/yes?

cc @abellotti @Fryguy 

Not sure if this deserved to be backported, but figured I would mark the same as #877