@Fryguy 2020-09-01
Is this a copy/similar to what we use in reporting?  include_for_find is literally the same thing it's called there, hence what triggered my mind - https://github.com/ManageIQ/manageiq/blob/69f14c83c53418b9549862e176934fd5afecefe5/app/models/miq_report/generator.rb#L339