@NickLaMuro 2020-09-01
I would say "no" simply because what is being worked on there is a `Hash`:

https://github.com/ManageIQ/manageiq/blob/b8e8f2e6250f4497ab44c695c1c5c4c1c53262e8/product/reports/400_Operations-%20Virtual%20Machines/010_Registered%20VMs%20by%20Free%20Space.yaml#L28-L35

And what we are working with here is a 1D `Array` with values that can be nested.  We would have to transform the data and then feed it into that.  Unsure if it is worth the effort, or a route we want to go down, but I don't really see it replacing much of what I have here.

Might be helpful with the `TODO` comment though.