@jrafanie 2020-09-02
Looks good so far, just asking questions to see if I'm understanding the changes that affect performance