@Fryguy 2020-09-09
As discussed, while support for the `:arel` clause is good, it's not complete, and we need support for the `:uses` clause as well.