@jrafanie 2020-09-10
> > Will un-`[WIP]` once we determine what to do with that last commit. (keep or drop)
> 
> I'm fine with keeping it.

@NickLaMuro I'm fine with it too.  👍  You don't seem happy with it so if you have good reasons to not include it, we can drop it, just let us know what you think.