@NickLaMuro 2020-09-11
> and we need support for the `:uses` clause as well.

So, this is going to be a bit rougher than I thought, and will probably take some time to think over how to implement (I have some half baked ideas, but nothing perfect at the moment).

I had to do a bit of gymnastics to make the "dot syntax" for relations to work, and probably have to do something different for `:uses` since it can be a collection of relations.  The `virtual_columns` on `VmOrTemplate` provide a good example of what we would have to try and support:

https://github.com/ManageIQ/manageiq/blob/9a23708582de38c9ae0cc8332f7e46fc1a53ba08/app/models/vm_or_template.rb#L139-L140

Because they do an `Array` of relations (`[:operating_system, :hardware]`), instead of say `"hardware.host.name"`, which expands to `{:hardware => {:host => {}}`, I am still trying to figure out how to pass this back through the `virtual_attributes_for` method.