@jrafanie 2020-09-16
@NickLaMuro I think I follow your last few commit changes. Thanks for explaining. Did you choose to not fix the remaining style issues?  The useless assignment seems legit.  The others are meh.

I think I'm good with this.  Are the performance numbers in the description still the same after the last commit?