@NickLaMuro 2020-09-16
> Did you choose to not fix the remaining style issues? The useless assignment seems legit. The others are meh.

One or two I agree with, so I will fix them an re-push in a bit.

> I think I'm good with this. Are the performance numbers in the description still the same after the last commit?

I hadn't re-run with the new changes, though said changes are also only expanding the types of attributes we support preloading for, and shouldn't affect the existing metrics.  Regardless, I think you bring up a good point that I should re-test to confirm all is well, so I will do that as well.