@NickLaMuro 2020-09-16
@jrafanie re-ran the **"After"** benchmarks I ran previously, and got _basically_ the same results:

```
$ bundle exec miqperf benchmark -ac 5 "/api/vms?expand=resources&attributes=name,hardware.cpu_sockets"
/api/vms
|   ms | queries | query (ms) | rows |
| ---: |    ---: |       ---: | ---: |
| 1757 |      12 |        140 | 2041 |
|  367 |      10 |       10.5 |  559 |
|  328 |      10 |        9.7 |  559 |
|  373 |      10 |       10.6 |  559 |
|  270 |      10 |        9.4 |  559 |
$ bundle exec miqperf benchmark -ac 5 "/api/vms?expand=resources,hardware&attributes=num_cpu,name"
/api/vms
|   ms | queries | query (ms) | rows |
| ---: |    ---: |       ---: | ---: |
| 1578 |      12 |         24 | 2041 |
|  330 |      10 |        7.3 |  559 |
|  351 |      10 |        8.3 |  559 |
|  205 |      10 |        8.1 |  559 |
|  198 |      10 |        7.3 |  559 |
```

Fixed the `rubocop` errors as well.