@simaishi 2020-09-24
Backported as a part of #890